# packages/api-contracts/openapi/chatdp.yaml
openapi: 3.0.3
info:
  title: ChatDP API
  version: 0.1.0
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:3000/api
    description: Local dev server
  - url: https://api.chatdp.duyphan-dz-dev.win/api
    description: Production server
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AuthTokens:
      type: object
      required: [accessToken, refreshToken, user]
      properties:
        accessToken:
          type: string
          description: JWT access token
        refreshToken:
          type: string
          description: JWT refresh token
        user:
          type: object
          properties:
            id:
              type: string
            email:
              type: string
              nullable: true
            displayName:
              type: string
            avatarUrl:
              type: string
              nullable: true
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
    SignupRequest:
      type: object
      required: [email, password, displayName]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        displayName:
          type: string
        avatarUrl:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    LogoutRequest:
      $ref: '#/components/schemas/RefreshRequest'
    MeResponse:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          nullable: true
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
        message:
          type: string
        error:
          type: string

    # Conversations
    Conversation:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [private, group]
        groupName:
          type: string
          nullable: true
        groupAvatarUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ConversationCreateRequest:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [private, group]
        participantUserIds:
          type: array
          items:
            type: string
          description: For private, must contain exactly one other user ID
        groupName:
          type: string
          nullable: true
        groupAvatarUrl:
          type: string
          nullable: true
    ConversationUpdateRequest:
      type: object
      properties:
        groupName:
          type: string
        groupAvatarUrl:
          type: string

    # Messages
    Message:
      type: object
      properties:
        id:
          type: string
        conversationId:
          type: string
        senderId:
          type: string
        contentType:
          type: string
          enum: [text, image, video, file, voice]
        content:
          type: string
          nullable: true
        mediaUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    MessageCreateRequest:
      type: object
      required: [contentType]
      properties:
        contentType:
          type: string
          enum: [text, image, video, file, voice]
        content:
          type: string
          nullable: true
        mediaUrl:
          type: string
          nullable: true
    MessageListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        nextCursor:
          type: string
          nullable: true
    MessageStatus:
      type: object
      properties:
        messageId:
          type: string
        userId:
          type: string
        status:
          type: string
          enum: [delivered, read]
        readAt:
          type: string
          format: date-time
          nullable: true
paths:
  /auth/signup:
    post:
      tags: [Auth]
      summary: Create a user account and issue tokens
      operationId: authSignup
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      operationId: authLogin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Exchange a valid refresh token for new tokens (rotation)
      operationId: authRefresh
      security: []
      description: |
        Issues a new access/refresh token pair and revokes the old refresh token.
        This endpoint may enforce binding of refresh tokens to the original User-Agent and/or client IP
        depending on server configuration. Clients should send a consistent `User-Agent` header and, when
        behind a proxy, preserve client IP via `X-Forwarded-For`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid or expired refresh token (or UA/IP mismatch when binding is enabled)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Revoke the current refresh token (idempotent)
      operationId: authLogout
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /me:
    get:
      tags: [Users]
      summary: Get current user profile
      operationId: usersMeGet
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          description: Unauthorized (missing/invalid JWT)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Conversations
  /conversations:
    get:
      tags: [Conversations]
      summary: List conversations for current user
      operationId: conversationsList
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Conversation'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [Conversations]
      summary: Create a conversation (private or group)
      operationId: conversationsCreate
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConversationCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /conversations/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
    get:
      tags: [Conversations]
      summary: Get conversation by id (must be a participant)
      operationId: conversationsGet
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags: [Conversations]
      summary: Update group conversation metadata
      operationId: conversationsUpdate
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConversationUpdateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /conversations/{id}/join:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
    post:
      tags: [Conversations]
      summary: Join a conversation
      operationId: conversationsJoin
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Messages
  /conversations/{conversationId}/messages:
    parameters:
      - in: path
        name: conversationId
        schema:
          type: string
        required: true
    get:
      tags: [Messages]
      summary: List messages in a conversation (seek pagination)
      operationId: messagesList
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - in: query
          name: cursor
          schema:
            type: string
          description: Message ID cursor (exclusive)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [Messages]
      summary: Create a message in the conversation
      operationId: messagesCreate
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /messages/{messageId}/read:
    parameters:
      - in: path
        name: messageId
        schema:
          type: string
        required: true
    post:
      tags: [Messages]
      summary: Mark a specific message as read by current user
      operationId: messagesRead
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Read status created/updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    $ref: '#/components/schemas/Message'
                  status:
                    $ref: '#/components/schemas/MessageStatus'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /messages/{messageId}/status:
    parameters:
      - in: path
        name: messageId
        schema:
          type: string
        required: true
    get:
      tags: [Messages]
      summary: List delivery/read statuses for a message (participants only)
      operationId: messagesStatuses
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MessageStatus'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'